EQ-100 Moved the extract function so it can be reused for all materialized view profiles.
