+++ etl/app/server/profiles/assessmentUnits.js
@@ -1,12 +1,7 @@
-import axios from 'axios';
-import crypto from 'crypto';
-import https from 'https';
-import { setTimeout } from 'timers/promises';
+import { extract as innerExtract } from './materializedViewsExtract.js';
 import pgPromise from 'pg-promise';
 
 const pgp = pgPromise({ capSQL: true });
-
-import { logger as log } from '../utilities/logger.js';
 
 /*
 ## Exports
@@ -65,34 +60,12 @@ const insertColumns = new pgp.helpers.ColumnSet([
 // Methods
 
 export async function extract(s3Config, next = 0, retryCount = 0) {
-  const chunkSize = s3Config.config.chunkSize;
-
-  const profileName = 'profile_assessment_units';
-
-  const url =
-    `${s3Config.services.materializedViews}/${profileName}` +
-    `?p_limit=${chunkSize}&p_offset=${next}`;
-
-  const res = await axios.get(url, {
-    headers: { 'API-key': '{887b8dd6-3cc5-4184-8c00-dc2b414f4a83}' },
-    httpsAgent: new https.Agent({
-      // TODO - Remove this when ordspub supports OpenSSL 3.0
-      // This is needed to allow node 18 to talk with ordspub, which does
-      //   not support OpenSSL 3.0
-      secureOptions: crypto.constants.SSL_OP_LEGACY_SERVER_CONNECT,
-    }),
-  });
-  if (res.status !== 200) {
-    log.info(`Non-200 response returned from ${profileName} service, retrying`);
-    if (retryCount < s3Config.config.retryLimit) {
-      await setTimeout(s3Config.config.retryIntervalSeconds * 1000);
-      return await extract(s3Config, next, retryCount + 1);
-    } else {
-      throw new Error('Retry count exceeded');
-    }
-  }
-  const data = res.data.records.map((record) => record);
-  return { data: data.length ? data : null, next: next + chunkSize };
+  return await innerExtract(
+    'profile_assessment_units',
+    s3Config,
+    next,
+    retryCount,
+  );
 }
 
 export function transform(data) {
